fix products tests
